AWS Integration and Lambda functions (#28)
* Renamed global queue to global bus

* Split up redis/ AWS queue abstractions

* More integration with S3/SQS

* Tweaks to tests

* Tests for sharing bus

* Failing integration test for SQS message bus

* Added base64 encoding of SQS messages to get round character restrictions

* Added JSON handling of messages

* Added json redis mode

* Factored out function loading/ compilation

* Isolated WAVM-related code

* Added reading/ writing files S3

* Added key deletion to S3 client

* Added AWS/ Pywren set-up instructions

* Added task to install AWS CPP runtime

* Added elasticache instance to mix

* Fixed upload of functions to S3

* Fixing aws teardown script

* Added scripts to download and build AWS lambda components

* Added lambda function target

* Renamed lambda target

* Skeleton of Lambda function upload

* Attempts to get Lambda function build working

* Changed lambda build to use main cmake files

* got lambda development loop working

* Hooked up C++ runtime properly

* Reducing verbosity

* Got echo function working on lambda

* Started pulling Redis functionality out into CoreRedis

* Fixed compile error

+++ lib/random.cpp
@@ -19,7 +19,7 @@ namespace faasm {
         std::shuffle(arrayIn, arrayIn + arrayLen, rng);
     }
 
-    int* randomIntRange(size_t rangeLen) {
+    int* randomIntRange(int rangeLen) {
         int *range = new int[rangeLen];
 
         for(int i = 0; i < rangeLen; i++) {
